% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouncermodels-package.R
\docType{package}
\name{bouncermodels-package}
\alias{bouncermodels}
\alias{bouncermodels-package}
\title{bouncermodels: Pre-Trained Models for Cricket Analytics}
\description{
Provides pre-trained machine learning models for cricket analytics, served via GitHub releases with local caching. Models include ball outcome prediction (agnostic and full), pre-match win probability, margin prediction, and score projection. Used by the bouncer package for match simulation and prediction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/peteowen1/bouncermodels}
  \item Report bugs at \url{https://github.com/peteowen1/bouncermodels/issues}
}

}
\author{
\strong{Maintainer}: Pete Owen \email{peteowen1@gmail.com}

}
\keyword{internal}
