% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{list_available_models}
\alias{list_available_models}
\title{List Available Models}
\usage{
list_available_models()
}
\value{
A list with three elements: ball_outcome, prediction, in_match
}
\description{
Returns a list of all available models that can be loaded.
}
\examples{
list_available_models()
}
