% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model_file}
\alias{load_model_file}
\title{Load a model file based on extension}
\usage{
load_model_file(path, label = basename(path))
}
\description{
Load a model file based on extension
}
\keyword{internal}
